Cope with no user on request
The Django test client's fake "login" code ends up calling the login signal handlers
without having set request.user to the user. That was resulting in an exception from
request-log that failed our tests that use login.

The user is passed to the login signal handler anyway, so just set it on the request
object if it's not there already.